requirements.txt
# Core dependencies
requests
python-dotenv
transformers
torch
sarvamai  # Sarvam AI SDK for Document Intelligence

# NLP and ML
datasets
seqeval
scikit-learn
numpy
pandas

# Visualization
matplotlib
seaborn

# Web interface
streamlit

# Testing
pytest
pytest-cov

# Development
black
flake8
ipython